need to rewrite with systematic approach to inputs

+++ console.h
@@ -4,7 +4,7 @@
 
 #include "source.h"
 
-void dump(const Source& source, int64_t n);
+void dump(const Source& source, size_t n);
 void dump_fm(size_t n);
 
 #endif
